removing the _Corvax folder and the moved content for cs
