+++ Content.Client/_Stories/DiscordAuth/DiscordAuthGui.xaml.cs
