+++ Content.Client/_Stories/DiscordAuth/DiscordAuthManager.cs
