+++ Content.Client/_Stories/JoinQueue/JoinQueueManager.cs
