+++ Content.Client/_Stories/JoinQueue/QueueGui.xaml.cs
