+++ Content.Client/_Stories/Sponsors/SponsorsManager.cs
