+++ Content.Server/_Stories/StationGoal/StationGoalCommand.cs
