+++ Content.Server/_Stories/StationGoal/StationGoalPaperComponent.cs
