+++ Content.Server/_Stories/StationGoal/StationGoalPaperSystem.cs
