+++ Content.Server/_Stories/StationGoal/StationGoalPrototype.cs
