+++ Content.Shared/_Stories/CCCVars/CCCVars.cs
@@ -1,14 +1,108 @@
-﻿// using Robust.Shared.Configuration;
+﻿using Robust.Shared.Configuration;
 
-// namespace Content.Shared._Stories.CCCVars;
+namespace Content.Shared._Corvax.CCCVars;
 
-// /// <summary>
-// ///     Stories modules console variables
-// /// </summary>
-// [CVarDefs]
-// // ReSharper disable once InconsistentNaming
-// public sealed class CCCVars
-// {
+/// <summary>
+///     Corvax modules console variables
+/// </summary>
+[CVarDefs]
+// ReSharper disable once InconsistentNaming
+public sealed class CCCVars
+{
+    /**
+     * TTS (Text-To-Speech)
+     */
 
+    /// <summary>
+    /// URL of the TTS server API.
+    /// </summary>
+    public static readonly CVarDef<bool> TTSEnabled =
+        CVarDef.Create("tts.enabled", false, CVar.SERVER | CVar.REPLICATED | CVar.ARCHIVE);
 
-// }
+    /// <summary>
+    /// URL of the TTS server API.
+    /// </summary>
+    public static readonly CVarDef<string> TTSApiUrl =
+        CVarDef.Create("tts.api_url", "", CVar.SERVERONLY | CVar.ARCHIVE);
+
+    /// <summary>
+    /// Auth token of the TTS server API.
+    /// </summary>
+    public static readonly CVarDef<string> TTSApiToken =
+        CVarDef.Create("tts.api_token", "", CVar.SERVERONLY | CVar.CONFIDENTIAL);
+
+    /// <summary>
+    /// Amount of seconds before timeout for API
+    /// </summary>
+    public static readonly CVarDef<int> TTSApiTimeout =
+        CVarDef.Create("tts.api_timeout", 5, CVar.SERVERONLY | CVar.ARCHIVE);
+
+    /// <summary>
+    /// Default volume setting of TTS sound
+    /// </summary>
+    public static readonly CVarDef<float> TTSVolume =
+        CVarDef.Create("tts.volume", 0f, CVar.CLIENTONLY | CVar.ARCHIVE);
+
+    /// <summary>
+    /// Count of in-memory cached tts voice lines.
+    /// </summary>
+    public static readonly CVarDef<int> TTSMaxCache =
+        CVarDef.Create("tts.max_cache", 250, CVar.SERVERONLY | CVar.ARCHIVE);
+
+    /*
+     * Sponsors
+     */
+
+    /// <summary>
+    ///     URL of the sponsors server API.
+    /// </summary>
+    public static readonly CVarDef<string> SponsorsApiUrl =
+        CVarDef.Create("sponsor.api_url", "", CVar.SERVERONLY);
+
+        public static readonly CVarDef<string> SponsorsDatabasePgHost =
+            CVarDef.Create("sponsor.pg_host", "localhost", CVar.SERVERONLY);
+
+        public static readonly CVarDef<int> SponsorsDatabasePgPort =
+            CVarDef.Create("sponsor.pg_port", 5432, CVar.SERVERONLY);
+
+        public static readonly CVarDef<string> SponsorsDatabasePgDatabase =
+            CVarDef.Create("sponsor.pg_database", "ss14", CVar.SERVERONLY);
+
+        public static readonly CVarDef<string> SponsorsDatabasePgUsername =
+            CVarDef.Create("sponsor.pg_username", "postgres", CVar.SERVERONLY);
+
+        public static readonly CVarDef<string> SponsorsDatabasePgPassword =
+            CVarDef.Create("sponsor.pg_password", "", CVar.SERVERONLY | CVar.CONFIDENTIAL);
+
+    /*
+     * Queue
+     */
+
+    /// <summary>
+    ///     Controls if the connections queue is enabled. If enabled stop kicking new players after `SoftMaxPlayers` cap and instead add them to queue.
+    /// </summary>
+    public static readonly CVarDef<bool>
+        QueueEnabled = CVarDef.Create("queue.enabled", false, CVar.SERVERONLY);
+
+    /*
+     * Discord Auth
+     */
+
+    /// <summary>
+    ///     Enabled Discord linking, show linking button and modal window
+    /// </summary>
+    public static readonly CVarDef<bool> DiscordAuthEnabled =
+        CVarDef.Create("discord_auth.enabled", false, CVar.SERVERONLY);
+
+    /// <summary>
+    ///     URL of the Discord auth server API
+    /// </summary>
+    public static readonly CVarDef<string> DiscordAuthApiUrl =
+        CVarDef.Create("discord_auth.api_url", "", CVar.SERVERONLY);
+
+    /// <summary>
+    ///     Secret key of the Discord auth server API
+    /// </summary>
+    public static readonly CVarDef<string> DiscordAuthApiKey =
+        CVarDef.Create("discord_auth.api_key", "", CVar.SERVERONLY | CVar.CONFIDENTIAL);
+}
